Fix MSVC compiler warnings
reg. conversion from 'size_t' to 'int', possible loss of data (C4267)

+++ src/project/internal/projectaudiosettings.cpp
@@ -366,7 +366,7 @@ QJsonObject ProjectAudioSettings::unitConfigToJson(const audio::AudioUnitConfig&
     QJsonObject result;
 
     for (const auto& pair : config) {
-        QByteArray byteArray = QByteArray::fromRawData(pair.second.c_str(), pair.second.size());
+        QByteArray byteArray = QByteArray::fromRawData(pair.second.c_str(), static_cast<int>(pair.second.size()));
         result.insert(QString::fromStdString(pair.first), QString(byteArray.toBase64()));
     }
